+++ python-lib/osisoft_constants.py
@@ -388,6 +388,6 @@ class OSIsoftConstants(object):
         "Security": "{base_url}/eventframes/{webid}/security",
         "SecurityEntries": "{base_url}/eventframes/{webid}/securityentries"
     }
-    PLUGIN_VERSION = "1.0.1-beta.6"
+    PLUGIN_VERSION = "1.0.1-beta.7"
     WEB_API_PATH = "piwebapi"
     WRITE_HEADERS = {'X-Requested-With': 'XmlHttpRequest'}