+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Disable by default server URL override
 - Fix pagination for Event Frames Search connector
 - Fix Transpose & Synchronize recipe with rows containing no timestamp
+- Fix nested value object when batch mode is not activated on Event Frames Search connector
 
 ## [Version 1.0.0](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.0.0) - Initial release - 2022-08-04